feat: 删除部门校验

+++ tlias-web-management/src/main/java/com/onedimension/service/impl/DeptServiceImpl.java
@@ -1,8 +1,12 @@
 package com.onedimension.service.impl;
 
 import com.onedimension.mapper.DeptMapper;
+import com.onedimension.mapper.EmpMapper;
 import com.onedimension.pojo.Dept;
+import com.onedimension.pojo.Emp;
+import com.onedimension.pojo.EmpQueryParams;
 import com.onedimension.pojo.Result;
+import com.onedimension.service.EmpService;
 import com.onedimension.utils.ResultUtil;
 import com.onedimension.service.DeptService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,6 +19,8 @@ public class DeptServiceImpl implements DeptService {
 
     @Autowired
     private DeptMapper deptMapper;
+    @Autowired
+    private EmpMapper empMapper;
 
     @Override
     public Result<List<Dept>> getDepts() {
@@ -28,21 +34,28 @@ public class DeptServiceImpl implements DeptService {
 
     @Override
     public Result deleteDept(Integer id) {
-        int rows = deptMapper.deleteDept(id);
-        if(rows > 0) {
-            return ResultUtil.success("删除成功");
+        EmpQueryParams empQueryParams = new EmpQueryParams();
+        empQueryParams.setDeptId(id);
+        List<Emp> page = empMapper.page(empQueryParams);
+        if (page.isEmpty()) {
+            int rows = deptMapper.deleteDept(id);
+            if (rows > 0) {
+                return ResultUtil.success("删除成功");
+            } else {
+                return ResultUtil.fail("删除失败, 部门不存在");
+            }
         } else {
-            return ResultUtil.fail("删除失败, 部门不存在");
+            throw new RuntimeException("删除失败, 部门下有员工");
         }
     }
 
     @Override
     public Result addDept(Dept dept) {
-        if(dept.getName() == null) {
+        if (dept.getName() == null) {
             return ResultUtil.fail("部门名不能为空");
         }
         int rows = deptMapper.addDept(dept);
-        if(rows > 0) {
+        if (rows > 0) {
             return ResultUtil.success("添加成功");
         } else {
             return ResultUtil.fail("添加失败");
@@ -52,7 +65,7 @@ public class DeptServiceImpl implements DeptService {
     @Override
     public Result<Dept> getDeptById(Integer id) {
         Dept dept = deptMapper.getDeptById(id);
-        if(dept != null) {
+        if (dept != null) {
             return ResultUtil.success(dept);
         } else {
             return ResultUtil.fail("部门不存在");
@@ -61,11 +74,11 @@ public class DeptServiceImpl implements DeptService {
 
     @Override
     public Result updateDept(Dept dept) {
-        if(dept.getName() == null || dept.getId() == null) {
+        if (dept.getName() == null || dept.getId() == null) {
             return ResultUtil.fail("部门名不能为空");
         }
         int rows = deptMapper.updateDept(dept);
-        if(rows > 0) {
+        if (rows > 0) {
             return ResultUtil.success("修改成功");
         } else {
             return ResultUtil.fail("修改失败");
